Updated scout-apm-php to depend on php-http/discovery to find PSR-17 and PSR-18 implementations
