+++ composer.json
@@ -11,6 +11,11 @@
         "ext-openssl": "*",
         "ext-sockets": "*",
         "ext-zlib": "*",
+        "php-http/discovery": "^1.14",
+        "psr/http-client": "^1.0",
+        "psr/http-client-implementation": "^1.0",
+        "psr/http-factory": "^1.0",
+        "psr/http-message": "^1.0",
         "psr/log": "^1.0|^2.0|^3.0",
         "psr/simple-cache": "^1.0",
         "ralouphie/getallheaders": "^2.0.5|^3.0",
@@ -21,6 +26,7 @@
         "composer-plugin-api": "^2.0",
         "cache/array-adapter": "^1.1",
         "doctrine/coding-standard": "^8.2",
+        "guzzlehttp/guzzle": "^7.3",
         "laravel/framework": "^5.5.0|^6.0|^7.0|^8.0",
         "laravel/lumen-framework": "^5.5.0|^6.0|^7.0|^8.0",
         "monolog/monolog": "^1.26|^2.2.0",
