+++ tests/Unit/Errors/ScoutClient/HttpErrorReportingClient.php
@@ -6,8 +6,8 @@ namespace Scoutapm\UnitTests\Errors\ScoutClient;
 
 use PHPUnit\Framework\TestCase;
 
-/** @covers \Scoutapm\Errors\ScoutClient\GuzzleErrorReportingClient */
-final class GuzzleErrorReportingClientTest extends TestCase
+/** @covers \Scoutapm\Errors\ScoutClient\HttpErrorReportingClient */
+final class HttpErrorReportingClient extends TestCase
 {
     public function testSendErrorToScout(): void
     {
